carry on

+++ models/file.go
@@ -7,7 +7,7 @@ type File struct {
 	Path     string  `json:"path" form:"path"`
 	Width    int     `json:"width" form:"width"`
 	Height   int     `json:"height" form:"height"`
-	Ratio    float32 `json:"ratio" form:"ratio"`
+	Ratio    float64 `json:"ratio" form:"ratio"`
 	Size     int64   `json:"size" form:"size"`
 	UserID   int64   `json:"user_id" form:"user_id"`
 }
